chore(model): refactor project structure
refactor project structure:
- validator package로 구분.
